+++ websites/common/config.example.inc.php
@@ -1,7 +1,7 @@
 <?php
 //This is an example file. Edit it accordingly and rename it to "config.inc.php"
 
-define('VERSION', '5.2.2');
+define('VERSION', '5.2.3');
 
 //Database access
 define('DB_HOST', 'YOUR_DB_HOST_HERE');